src/main/java/org/example/pageobject/AccountPage.java
package org.example.pageobject;

import org.openqa.selenium.By;
import org.openqa.selenium.WebDriver;
import org.openqa.selenium.support.ui.ExpectedConditions;
import org.openqa.selenium.support.ui.WebDriverWait;
import java.time.Duration;


public class AccountPage {
    private WebDriver driver;
    private By accountMessage = By.xpath(
            "//p[text()='В этом разделе вы можете изменить свои персональные данные']");
    private By exitButton = By.xpath("//button[text()='Выход']");

    public AccountPage(WebDriver driver) {
        this.driver = driver;
    }

    public boolean isAccountMessageVisible() {
        new WebDriverWait(driver, Duration.ofSeconds(3))
                .until(ExpectedConditions.visibilityOfElementLocated(accountMessage));
        return driver.findElement(accountMessage).isDisplayed();
    }

    public void clickExitButton() {
        new WebDriverWait(driver, Duration.ofSeconds(3))
                .until(ExpectedConditions.visibilityOfElementLocated(exitButton));
        driver.findElement(exitButton).click();
    }
}